Use zerolog for logging
Closes #6

This will switch the default logging format to JSON. For simplicity, it
will also change from multiple `v` arguments for increased verbosity to
a single `-d` or `--debug` flag. A side effect of this is that `-v` can
again be used for version. It seems it was being overloaded previously.